fix(voice-agents): serialize total_credits as a number, not a string
VoiceAgentResponse.total_credits was typed as Decimal, which Pydantic
serializes to a JSON string. Every sibling response DTO (chat_session,
client_session, execution) exposes the same Numeric column as float,
so this broke frontend arithmetic on totalCredits. The ORM column
stays Decimal; only the response DTO's wire type changes.

+++ assemblix-app-api/assemblix_api/dto/responses/voice_agent.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import Field
@@ -20,6 +19,6 @@ class VoiceAgentResponse(DTOModel):
     config: VoiceAgentConfig = Field(description="Prompt, voice, knowledge and analysis hooks")
     is_active: bool = Field(description="Whether the agent can accept sessions")
     session_count: int = Field(description="Number of voice sessions started with this agent")
-    total_credits: Decimal = Field(description="Total credits consumed by this agent")
+    total_credits: float = Field(description="Total credits consumed by this agent")
     created_at: datetime = Field(description="Creation timestamp")
     updated_at: datetime = Field(description="Last update timestamp")
